Fixed elastica equation test (float64 and bc without penalty).

+++ tests/conftest.py
@@ -16,5 +16,6 @@ import dctkit
 def setup_test():
     # NOTE: running multiple JAX tests with different data types DOES NOT work
     # (jax_enable_x64 must be changed AT STARTUP)
+    # WARNING: test_elastica fails with float32 precision
     dctkit.config_called = False
-    config(FloatDtype.float32, IntDtype.int32, Backend.jax, Platform.cpu)
+    config(FloatDtype.float64, IntDtype.int64, Backend.jax, Platform.cpu)
